+++ volcengine-java-sdk-seccenter20240508/src/main/java/com/volcengine/seccenter20240508/model/DataForGetDevFingerprintPortOutput.java
@@ -67,7 +67,7 @@ public class DataForGetDevFingerprintPortOutput {
   private String sport = null;
 
   @SerializedName("StartTime")
-  private Integer startTime = null;
+  private Long startTime = null;
 
   @SerializedName("Status")
   private String status = null;
@@ -76,7 +76,7 @@ public class DataForGetDevFingerprintPortOutput {
   private String uid = null;
 
   @SerializedName("UpdateTime")
-  private Integer updateTime = null;
+  private Long updateTime = null;
 
   @SerializedName("Username")
   private String username = null;
@@ -297,7 +297,7 @@ public class DataForGetDevFingerprintPortOutput {
     this.sport = sport;
   }
 
-  public DataForGetDevFingerprintPortOutput startTime(Integer startTime) {
+  public DataForGetDevFingerprintPortOutput startTime(Long startTime) {
     this.startTime = startTime;
     return this;
   }
@@ -307,11 +307,11 @@ public class DataForGetDevFingerprintPortOutput {
    * @return startTime
   **/
   @Schema(description = "")
-  public Integer getStartTime() {
+  public Long getStartTime() {
     return startTime;
   }
 
-  public void setStartTime(Integer startTime) {
+  public void setStartTime(Long startTime) {
     this.startTime = startTime;
   }
 
@@ -351,7 +351,7 @@ public class DataForGetDevFingerprintPortOutput {
     this.uid = uid;
   }
 
-  public DataForGetDevFingerprintPortOutput updateTime(Integer updateTime) {
+  public DataForGetDevFingerprintPortOutput updateTime(Long updateTime) {
     this.updateTime = updateTime;
     return this;
   }
@@ -361,11 +361,11 @@ public class DataForGetDevFingerprintPortOutput {
    * @return updateTime
   **/
   @Schema(description = "")
-  public Integer getUpdateTime() {
+  public Long getUpdateTime() {
     return updateTime;
   }
 
-  public void setUpdateTime(Integer updateTime) {
+  public void setUpdateTime(Long updateTime) {
     this.updateTime = updateTime;
   }
